improve layout, add model selection
