+++ README.md
@@ -19,6 +19,7 @@ poetry install
 
 ## Usage
 streamlit run app.py
+Before that you should specify a Redis database by `conda env config vars set REDIS_OM_URL={YOUR_URL}`
 
 ### Chat
 First choose the agents (two agents cannot be the same), scenarios and the agent you are going to be, then click `start` to start interaction. When you want to leave and get evaluated, click `stop` to start evaluation.
