+++ socialstream/chat/__init__.py
@@ -1,3 +1,3 @@
-from .chat_new import chat_demo
+from .chat import chat_demo
 
 __all__ = ["chat_demo"]
